% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa.R
\name{draw_key_usa}
\alias{draw_key_usa}
\title{Title}
\usage{
draw_key_usa(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point(key_glyph = draw_key_usa)
}
